+++ lib/type.js
@@ -1,0 +1,3 @@
+module.exports = function Type(signature, name) {
+	return { type: signature, name: name };
+};
